Create rds.tf
aws_db_instance: This resource is used to create an RDS MySQL instance with the specified configuration options. The following parameters are set:
allocated_storage: The amount of storage to be allocated to the instance in GB.
db_name: The name of the database to be created on the instance.
engine: The database engine to use. In this case, it is MySQL.
engine_version: The version of the database engine to use. In this case, it is MySQL version 5.7.
instance_class: The instance type to be used for the RDS instance.
username and password: The credentials to be used to access the instance. These are sensitive variables that will be provided in Terraform Cloud.
vpc_security_group_ids: The security group to be attached to the instance.
db_subnet_group_name: The name of the subnet group for the instance.
skip_final_snapshot: Whether or not to create a final snapshot when the instance is deleted.
aws_db_subnet_group: This resource is used to create a subnet group for the RDS instance. The following parameters are set:
name: The name of the subnet group.
subnet_ids: The IDs of the subnets in which the RDS instance will be launched.
tags: Tags to be attached to the subnet group.